contract-sdk: add macro to support single-crate contracts
