+++ Cargo.toml
@@ -14,7 +14,6 @@ members = [
     "contract-sdk/types",
     "contract-sdk/storage",
     "contract-sdk-macros",
-    "contract-sdk/specs/oas20/types",
 
     # Test runtimes.
     "tests/runtimes/benchmarking",
